backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
numpy
google-cloud-storage
firebase-admin
llama_index
openai